+++ ChangeLog.md
@@ -3,6 +3,11 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
+* 2017/07/06 - 4.14
+    * Allowed read-only role to see new concept view screen
+    * Fixed error message display on concept properties
+    * Fixed passing of terminology types when creating a child concept
+
 * 2017/06/29 - 4.13
     * Added UUID to Designations
     * added STAMP to concept, designations, properties (except VUID and Code), and relationships